requirements.txt
asgiref==3.5.0
aspy.refactor-imports==3.0.1
astor==0.8.1
attrs==21.4.0
autoflake==1.4
autopep8==1.6.0
bandit==1.7.4
black==22.6.0
click==8.1.3
darglint==1.8.1
Django==3.2
django-debug-toolbar==3.2.4
django-split-settings==1.1.0
docutils==0.18.1
eradicate==2.1.0
Faker==13.3.4
flake8==4.0.0
flake8-bandit==3.0.0
flake8-broken-line==0.4.0
flake8-bugbear==22.3.23
flake8-commas==2.1.0
flake8-comprehensions==3.8.0
flake8-debugger==4.0.0
flake8-docstrings==1.6.0
flake8-eradicate==1.2.0
flake8-isort==4.1.1
flake8-polyfill==1.0.2
Flake8-pyproject==0.9.0
flake8-quotes==3.3.1
flake8-rst-docstrings==0.2.5
flake8-string-format==0.3.0
gitdb==4.0.9
GitPython==3.1.27
iniconfig==1.1.1
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pbr==5.8.1
pep8-naming==0.12.1
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9
py==1.11.0
pycodestyle==2.8.0
pydantic==1.9.1
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
restructuredtext-lint==1.4.0
seed-isort-config==2.2.0
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
sqlparse==0.4.2
stevedore==3.5.0
testfixtures==6.18.5
toml==0.10.2
tomli==2.0.1
typing-extensions==4.2.0
wemake-python-styleguide==0.16.1
